Separates sections from main page and adds send contact email
