+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "lucide-react": "^0.507.0",
     "next": "15.2.4",
+    "nodemailer": "^7.0.3",
     "postcss": "^8.5.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -20,6 +21,7 @@
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
+    "@types/nodemailer": "^6.4.17",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
