+++ src/app/Footer.tsx
@@ -1,0 +1,159 @@
+"use client";
+import { useTranslations } from "../context/translations/TranslationsProvider";
+import Link from "next/link";
+
+const Footer = () => {
+  const { translations: t } = useTranslations();
+  return (
+    <footer className="border-t py-12 md:py-16">
+      <div className="container">
+        <div className="grid gap-8 sm:grid-cols-2 md:grid-cols-4">
+          <div>
+            <div className="flex items-center gap-2 font-bold text-xl mb-4">
+              <span className="text-primary">TUSK</span> Trade Company
+            </div>
+            <p className="text-sm text-muted-foreground">{t.footerText}</p>
+          </div>
+
+          <div>
+            <h3 className="font-semibold mb-4">{t.productsFooter}</h3>
+            <ul className="space-y-2 text-sm">
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.beef}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.lamb}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.poultry}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.dairy}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  Horse Meat
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.grains}
+                </Link>
+              </li>
+            </ul>
+          </div>
+
+          <div>
+            <h3 className="font-semibold mb-4">{t.companyFooter}</h3>
+            <ul className="space-y-2 text-sm">
+              <li>
+                <Link
+                  href="#about"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.aboutUs}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#global"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.globalReachFooter}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.certifications}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.sustainability}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#contact"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.contact}
+                </Link>
+              </li>
+            </ul>
+          </div>
+
+          <div>
+            <h3 className="font-semibold mb-4">{t.legal}</h3>
+            <ul className="space-y-2 text-sm">
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.terms}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.privacy}
+                </Link>
+              </li>
+              <li>
+                <Link
+                  href="#"
+                  className="text-muted-foreground hover:text-foreground"
+                >
+                  {t.cookies}
+                </Link>
+              </li>
+            </ul>
+          </div>
+        </div>
+
+        <div className="mt-12 pt-8 border-t text-center text-sm text-muted-foreground">
+          <p>
+            © {new Date().getFullYear()} TUSK Trade Company. {t.rights}
+          </p>
+        </div>
+      </div>
+    </footer>
+  );
+};
+
+export default Footer;
