+++ src/app/Global.tsx
@@ -1,0 +1,98 @@
+import { Globe, MapPin } from "lucide-react";
+import Image from "next/image";
+import { useTranslations } from "../context/translations/TranslationsProvider";
+
+const Global = () => {
+  const { translations: t } = useTranslations();
+  return (
+    <section id="global" className="py-16 md:py-24 bg-muted/50">
+      <div className="container">
+        <div className="text-center max-w-3xl mx-auto mb-16">
+          <h2 className="text-3xl font-bold tracking-tight mb-4">
+            {t.globalTitle}
+          </h2>
+          <p className="text-lg text-muted-foreground">{t.globalSubtitle}</p>
+        </div>
+
+        <div className="grid md:grid-cols-2 gap-12 items-center">
+          <div className="relative h-[400px] rounded-lg overflow-hidden">
+            <Image
+              src="/map.png"
+              alt="World map showing global trade routes"
+              fill
+              className="object-contain"
+            />
+          </div>
+
+          <div>
+            <div className="mb-8">
+              <div className="flex items-center gap-3 mb-4">
+                <MapPin className="h-5 w-5 text-primary" />
+                <h3 className="text-xl font-bold">{t.supplierCountries}</h3>
+              </div>
+              <p className="text-muted-foreground mb-4">{t.supplierText}</p>
+              <div className="grid grid-cols-2 gap-2">
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>Uruguay</span>
+                </div>
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>Paraguay</span>
+                </div>
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>Brazil</span>
+                </div>
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>Argentina</span>
+                </div>
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>Chile</span>
+                </div>
+              </div>
+            </div>
+
+            <div>
+              <div className="flex items-center gap-3 mb-4">
+                <Globe className="h-5 w-5 text-primary" />
+                <h3 className="text-xl font-bold">{t.clientMarkets}</h3>
+              </div>
+              <p className="text-muted-foreground mb-4">{t.clientText}</p>
+              <div className="grid grid-cols-2 gap-2">
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>Russian Federation</span>
+                </div>
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>United Kingdom</span>
+                </div>
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>Germany</span>
+                </div>
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>China</span>
+                </div>
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>United States</span>
+                </div>
+                <div className="flex items-center gap-2">
+                  <div className="h-2 w-2 rounded-full bg-primary"></div>
+                  <span>{t.andMore}</span>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default Global;
