+++ src/app/api/email/route.ts
@@ -1,0 +1,40 @@
+import { type NextRequest, NextResponse } from "next/server";
+import nodemailer from "nodemailer";
+import Mail from "nodemailer/lib/mailer";
+
+export async function POST(request: NextRequest) {
+  const { email, name, message } = await request.json();
+
+  const transport = nodemailer.createTransport({
+    service: "gmail",
+    auth: {
+      user: process.env.email,
+      pass: process.env.password,
+    },
+  });
+
+  const mailOptions: Mail.Options = {
+    from: process.env.email,
+    to: process.env.email,
+    subject: `Message from ${name} (${email})`,
+    text: message,
+  };
+
+  const sendMailPromise = () =>
+    new Promise<void>((resolve, reject) => {
+      transport.sendMail(mailOptions, function (err) {
+        if (!err) {
+          resolve();
+        } else {
+          reject(err.message);
+        }
+      });
+    });
+
+  try {
+    await sendMailPromise();
+    return NextResponse.json({ message: "Email sent" });
+  } catch (err) {
+    return NextResponse.json({ error: err }, { status: 500 });
+  }
+}
