+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import TranslationsProvider from "../context/translations/TranslationsProvider";
+import Footer from "./Footer";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -28,7 +29,10 @@ export default function RootLayout({
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        <TranslationsProvider>{children}</TranslationsProvider>
+        <TranslationsProvider>
+          {children}
+          <Footer />
+        </TranslationsProvider>
       </body>
     </html>
   );
