+++ src/app/sendEmail.ts
@@ -1,0 +1,13 @@
+interface EmailData {
+  email: string;
+  name: string;
+  message: string;
+}
+
+export async function sendEmail(data: EmailData): Promise<Response> {
+  const apiEndpoint = "/api/email";
+  return fetch(apiEndpoint, {
+    method: "POST",
+    body: JSON.stringify(data),
+  });
+}
